Refactoring project structure - 2 main modules created (core and features) - ready for new feature implementations
