+++ singular-web/src/app/app.routes.ts
@@ -10,34 +10,36 @@ export const routes: Routes = [
   },
   {
     path: 'login',
-    loadComponent: () => import('./auth/login/login.component').then((m) => m.LoginComponent),
+    loadComponent: () =>
+      import('./features/auth/login/login.component').then((m) => m.LoginComponent),
     canActivate: [inverseAuthGuard],
   },
   {
     path: 'register',
     loadComponent: () =>
-      import('./auth/register/register.component').then((m) => m.RegisterComponent),
+      import('./features/auth/register/register.component').then((m) => m.RegisterComponent),
     canActivate: [inverseAuthGuard],
   },
   {
-    path: ':id',
-    loadComponent: () => import('./layout/root/root.component').then((m) => m.RootComponent),
+    path: ':userId',
+    loadComponent: () => import('./core/layout/root/root.component').then((m) => m.RootComponent),
     canActivate: [authGuard],
     children: [
       { path: '', pathMatch: 'full', redirectTo: 'home' },
       {
         path: 'home',
-        loadComponent: () => import('./pages/home/home.compontent').then((m) => m.HomeComponent),
+        loadComponent: () =>
+          import('./features/profile/home/home.compontent').then((m) => m.HomeComponent),
       },
       {
         path: 'profile',
         loadComponent: () =>
-          import('./pages/profile/profile.component').then((m) => m.ProfileComponent),
+          import('./features/profile/profile/profile.component').then((m) => m.ProfileComponent),
       },
       {
         path: 'profile/edit',
         loadComponent: () =>
-          import('./pages/edit-personal-info/edit-personal-info.component').then(
+          import('./features/profile/edit-personal-info/edit-personal-info.component').then(
             (m) => m.EditPersonalInfoComponent,
           ),
       },
@@ -46,6 +48,6 @@ export const routes: Routes = [
   {
     path: '**',
     loadComponent: () =>
-      import('./pages/not-found/not-found.component').then((m) => m.NotFoundComponent),
+      import('./core/pages/not-found/not-found.component').then((m) => m.NotFoundComponent),
   },
 ];
