+++ singular-web/src/app/core/guards/auth.guard.ts
@@ -2,7 +2,7 @@ import { CanActivateFn, Router } from '@angular/router';
 import { inject } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { take, tap } from 'rxjs';
-import { selectIsAuthed } from '../../auth/state';
+import { selectIsAuthed } from '../../features/auth/state';
 
 export const authGuard: CanActivateFn = () => {
   const store = inject(Store);
