+++ singular-web/src/app/core/layout/root/root.component.html
