+++ singular-web/src/app/features/auth/login/login.component.scss
