+++ singular-web/src/app/features/auth/register/register.component.scss
