+++ singular-web/src/app/features/auth/state/auth.selectors.ts
