+++ singular-web/src/app/features/profile/profile/profile.component.scss
